kernel: add new ltq-vdsl-vr11-mei driver
This uses version 1.9.3 of the drv_mei_cpe package from the Intel UGW
8.3.2.30 for the VRX518.

This version maybe also works for the vr9 target, but the last
official supported version for vr9 is (AFAIK) 1.5.17.7 in the
"DSL Subsystem Release 5.7.6.2 (R7) for VRX200".
So, let's stay with version 1.5.17.6 for vr9 at the moment.

Signed-off-by: Martin Schiller <ms.3headeddevs@gmail.com>
[updated for kernel 5.10]
Signed-off-by: Jan Hoffmann <jan@3e8.eu>