kernel: add new ltq-vdsl-vr11 driver
This uses version 4.21.3 of the drv_dsl_cpe_api package from the Intel
UGW 8.3.2.30 for the VRX518.

This version maybe also works for the vr9 target, but the last
official supported version for vr9 is (AFAIK) 4.17.18.7 in the
"DSL Subsystem Release 5.7.6.2 (R7) for VRX200".
So, let's stay with version 4.17.18.6 for vr9 at the moment.

Signed-off-by: Martin Schiller <ms.3headeddevs@gmail.com>