ltq-vectoring: add driver
This driver provides a callback for sending vectoring error samples.

It is based on the open source code released for FRITZ!Box 7430 7.27.
Changes were made to support more recent kernel versions and the used
interface was changed from ptm0 to dsl0. The procfs path has been
cleaned up and a few missing fields in the SKBs are filled.