ltq-vdsl-mei: use polling instead of interrupts
For some users, unexpected resyncs happen at random intervals (usually
every few days). It seems that this is caused by concurrent read and
write access to the mailbox. Switching to polling makes this problem go
away, as the call to MEI_ProcessIntPerVrxLine (which eventually reads
messages from the mailbox) is protected by a lock in this case.

While this issue is not really specific to vectoring, the additional
interrupts for error reports make it much more likely to actually
trigger the issue.

Signed-off-by: Jan Hoffmann <jan@3e8.eu>