realtek: add support for HPE 1920 series
This adds support for the following models:
- HPE 1920-8G (untested)
- HPE 1920-16G
- HPE 1920-24G (untested, but same board as 16G)
- HPE 1920-48G (untested)

Booting initramfs image and initial installation:

- Prepare a FTP or TFTP server serving the OpenWrt initramfs image and
connect the server to a switch port

- Connect to the console port of the device and enter the extended
boot menu by typing Ctrl+B when prompted

- Choose the menu option <3> Enter Ethernet SubMenu

- Set network parameters via the option <5> Modify Ethernet Parameter,
enter FTP/TFTP filename as "Load File Name", the "Target File Name"
is not required for booting from RAM
(this only needs to be done once, the settings are saved)

- Select <1> Download Application Program To SDRAM And Run

- Once OpenWrt has booted it can be installed permanently via sysupgrade

- After initial installation, the bootloader needs to be configured to
load the correct image file

- Enter the extended boot menu again and choose <4> File Control

- Select <2> Set Application File type

- Enter the number of "openwrt-kernel.bin" (should be 1)

- Select the option <1> +Main to set the image as boot image

- Choose <0> Exit To Main Menu and then <1> Boot System

- OpenWrt should now be booting from flash!

NOTE: The bootloader on these can only boot from the VFS filesystem
which normally spans most of the flash. With OpenWrt, only the first
part of the firmware partition contains a valid filesystem, the rest is
used for rootfs. As the bootloader does not know about this, you must
not do any file operations in the bootloader, as this may corrupt the
OpenWrt installation (selecting the boot image is an exception, as it
only stores a flag in the bootloader data, but doesn't write to the
filesystem).

Signed-off-by: Jan Hoffmann <jan@3e8.eu>