firmware-utils: add image and filesystem tools for H3C devices
mkh3cimg creates a firmware image. Firmware images for these devices can
contain multiple files, but this tool is limited to creating images with
a single application file. In the case of OpenWrt, this is going to
contain the kernel image. Compressed files are supported by the image
format, in this case the supplied input file needs to a 7z archive with
LZMA compression, with unencoded header.

mkh3cvfs creates a VFS filesystem image. The bootloader can only boot
images from this filesystem which normally spans most of the flash on
the device. This tool creates a filesystem image containing a single
file. While the size of the filesystem is hardcoded in the bootloader,
an image that is only slightly larger than its content is sufficient,
as long as no write operations are to be performed on it. This allows to
append further data (i.e. rootfs) after the end of the generated image.

Signed-off-by: Jan Hoffmann <jan@3e8.eu>