realtek: add cond_resched to rtl838x_smi_wait_op
During PHY configuration lots of MDIO writes are performed in a loop.
If the kernel is built without preemption, this can prevent other kernel
tasks from being scheduled for more than a second. On a device with an
external watchdog, this can result in an unwanted reboot. Fix this by
adding a cond_resched call to the wait loop in rtl838x_smi_wait_op.

Signed-off-by: Jan Hoffmann <jan@3e8.eu>