realtek: restructure rtl_table_read/write
These two functions are identical apart from writing different values to
the read/write bit. Create a new function rtl_table_exec to reduce code
duplication.

Also replace the unbounded busy-waiting loop. As the hardware usually
responds very quickly, always call cond_resched, so that callers do not
need to worry about blocking for too long. This is especially important
for the L2 table, as a full loop (as in rtl83xx_port_fdb_dump) takes
about 20ms on RTL839x (and that function actually ends up being called
in a loop for every port).

So far, polling timeout errors are only handled by logging an error, but
a return value is added to allow proper handling in the future.

Signed-off-by: Jan Hoffmann <jan@3e8.eu>