realtek: don't treat first multicast portmask entry as reserved
There doesn't actually appear to be a reason to do this, only the last
entry is actually reserved for unknown multicast.

This also fixes two issues: As the increment happened after the bounds
check, the last reserved value could be overwritten. On deletion of
entries, a corresponding decrement was missing, causing the wrong entry
to be marked as free.

Fixes: cde31976e375 ("realtek: Add support for Layer 2 Multicast")
Signed-off-by: Jan Hoffmann <jan@3e8.eu>