realtek: don't set L2LEARNING flag in rtl83xx TX header
Enabling L2 learning for packets forwarded by the CPU can result in
incorrect entries being written to the forwarding table.

If a packet that was trapped to the CPU port gets forwarded to another
port, the switch incorrectly learns the CPU port instead of the original
source port. If the packet is sent untagged, this entry is also created
for rvid 0 instead of the real VLAN.

Examples for packets which trigger this issue are Multicast Listener
Reports and IGMP membership reports.

On RTL839x, this change doesn't have any effect, as learning on the CPU
port appears to be disabled by default in RTL839X_L2_PORT_NEW_SALRN
(this is not explicitly configured by the driver).

On RTL838x, this may cause some traffic for the CPU port to be flooded.
However, this inefficiency is much less of an issue than misforwarding
traffic to the switch CPU every time a connected device sends an IGMP or
MLD report.

The severity of flooding is also somewhat limited, as the eth0 address
does not need an entry in the forwarding table (it is written to a
special MAC address register in the switch).

Fixes: 9eab76c84e31 ("realtek: Improve TX CPU-Tag usage")
Signed-off-by: Jan Hoffmann <jan@3e8.eu>