WIP: ipq40xx: emulate Lantiq SoC in PCIe driver
This is a hack that is included in vendor firmware (taken from Fritzbox
7530 open source archive), modified to apply to the upstream driver.

It sets the PCIe vendor and device ID to the ones of the Lantiq GRX500
PCIe host. The address range providing chip identification on Lantiq
SoCs is remapped to a buffer containing the manufacturer ID of Lantiq
using the address translation unit.

Signed-off-by: Jan Hoffmann <jan@3e8.eu>