realtek: actually initialize and use num_lag_ids
The field num_lag_ids in the dsa_switch struct is not initialized, but
still used in rtl83xx_lag_can_offload. As a result, the check always
returns false and the LAG won't be offloaded.

Initialize it, and also fix the check to take into account that LAG IDs
managed by DSA are 1-based.

While at it, replace all uses of the n_lags field in the switch private
struct by the num_lag_ids field, to avoid duplication.

TODO: Fixes
Signed-off-by: Jan Hoffmann <jan@3e8.eu>