realtek: call rtl9300_serdes_setup for additional PHY interface modes
This is a preparation to properly support the RTL8826 PHY, which
requires switching between SGMII and HSGMII to support all link speeds.

Ideally, we should be able to properly configure all modes (also to stop
relying on intial setup by the bootloader). However, including XGMII
mode breaks networking for the RTL8218D on Zyxel XGS1010-12, so keep
that a special case for now.

Signed-off-by: Jan Hoffmann <jan@3e8.eu>