realtek: improve RTL8226 support
Refactor the driver for the RTL8226 PHY based on upstream drivers for
similar PHYs from recent kernels.

This fixes various issues with this PHY. The most important one is
switching between SGMII and HSGMII mode, which is required to support
all links speeds (the PHY uses HSGMII for 2.5G, and sgmii for everything
else).

Other fixes include proper configuration and status reporting for auto-
negotiation (including all supported link speeds), as well as working
suspend/resume.

The same changes are applied to RTL8221B, but this is untested.

Signed-off-by: Jan Hoffmann <jan@3e8.eu>