realtek: fix coexistence of 1G and 2.5G modes on RTL930x SerDes pair
Currently, it is not possible for a 1G and a 2.5G PHY mode to coexist on
a pair of SerDes on RTL930x. When trying to configure it like that, only
one of the port works.

Looking at how the parameters lc_on and lc_value are configured in
rtl9300_force_sds_mode, it becomes clear why this is the case. These
parameters are written to lane 0 of the SerDes pair, in register 18 on
page 0x20. This register has the following fields:

- bits 15-12: lc_value that is used for lc_on=false
- bits 11-8: lc_value that is used for lc_on=true
- bits 7-6: lane 1 lc_on (0x3 if true, else 0x1)
- bits 5-4: lane 0 lc_on (0x3 if true, else 0x1)
- bits 3-0: enable bits (not relevant here)

lc_value is specific to the speed (1G = 0x1, 2.5G = 0x3, 10G = 0x5),
and lc_on currently has a fixed value for each mode. Specifically, both
1G and 2.5G modes use lc_on=false. So, if both are configured on the
two lanes of a SerDes pair, both use the lc_value from bits 15-12. But
since the field can obviously only take one value, this can't work.

To actually allow both lanes of a SerDes to operate at different speeds,
it is necessary that both use a different lc_on value. However, not
every speed works with both lc_on values.

This implementation chooses to configure lc_on dynamically for 1G modes.
That approach works fine on Zyxel XGS1010-12. Choosing lc_on dynamically
for 2.5G modes, like the original code comments suggest, does not work
on this device, as the 2.5G modes only work with lc_on=false.

However, the vendor firmware for the very similar Zyxel XGS1210-12
actually uses lc_on=true for HSGMII. So it is not quite clear what is
going on here: There might be different hardware revisions, or maybe it
is just dependent on previous configuration (e.g. by the bootloader).
So maybe this needs to be revisited in the future.

While at it, remove all the debug output that is specific for SDS 5.

Signed-off-by: Jan Hoffmann <jan@3e8.eu>