realtek: fix ethtool stats for RTL839x and RTL930x
The MIB registers contain different stats depending on the SoC, and for
RTL930x some stats are in an additional register.

While we're at it also make reading 64-bit counters more robust, by
protecting against an overflow of the lower 32 bits during the read.

RTL931x remains unsupported, because it uses a table and thus requires
a separate implementation.

Signed-off-by: Jan Hoffmann <jan@3e8.eu>