realtek: support configuring SerDes auto-negotiation on RTL930x
There are SFP modules which only work if auto-negotiation is disabled,
like some "OEM SFP-2.5G-T" modules. This also seems to be necessary for
RTL8226/RTL8221B PHYs when using 2500Base-X.

However, currently, it is always enabled, so add support for configuring
it to make these SFP modules and PHYs work.

This implementation calls directly into functions of the MDIO driver,
because this is currently the only way to access the SerDes for ports
with a PHY.

Signed-off-by: Jan Hoffmann <jan@3e8.eu>